+++ plot/viewer/src/canvases/FoundationCanvas.tsx
@@ -18,7 +18,7 @@ export function FoundationCanvas(props: SketchCanvasProps) {
       showFoldButton={false}
       injectAnchor={true}
       applyAnchorRadialLayout={true}
-      enableAutoLayout={true}
+      layoutAlgo="tree"
     />
   );
 }