+++ plot/viewer/src/canvases/ServicesCanvas.tsx
@@ -17,6 +17,7 @@ export function ServicesCanvas(props: SketchCanvasProps) {
       shouldDrill={shouldDrillService}
       showFoldButton={true}
       injectAnchor={true}
+      layoutAlgo="radial"
     />
   );
 }